# Enhance joining and grouping