--- a/REQUIRE
+++ b/REQUIRE
@@ -4,5 +4,5 @@ StatsBase 0.8.3
 GZip
 SortingAlgorithms
 Reexport
-Compat 0.8
+Compat 0.8.4
 FileIO 0.1.2