--- a/REQUIRE
+++ b/REQUIRE
@@ -1,7 +1,7 @@
 julia 0.5
 NullableArrays 0.0.8
 CategoricalArrays 0.0.6
-StatsBase 0.8.3
+StatsBase 0.11.0
 GZip
 SortingAlgorithms
 Reexport