--- a/REQUIRE
+++ b/REQUIRE
@@ -7,3 +7,4 @@ SortingAlgorithms
 Reexport
 Compat 0.8.4
 FileIO 0.1.2
+Juno 0.2.4