--- a/docs/make.jl
+++ b/docs/make.jl
@@ -6,7 +6,7 @@ using Documenter, DataFrames
 makedocs(
     # options
     modules = [DataFrames],
-    doctest = false,
+    doctest = true,
     clean = false,
     sitename = "DataFrames.jl",
     format = Documenter.Formats.HTML,