--- a/src/DataFrames.jl
+++ b/src/DataFrames.jl
@@ -92,6 +92,12 @@ export @~,
 ##
 ##############################################################################
 
+if VERSION < v"0.5.0-dev+2023"
+    _displaysize(x...) = Base.tty_size()
+else
+    const _displaysize = Base.displaysize
+end
+
 for (dir, filename) in [
         ("other", "utils.jl"),
         ("other", "index.jl"),