--- a/src/DataFrames.jl
+++ b/src/DataFrames.jl
@@ -113,6 +113,7 @@ for (dir, filename) in [
         ("subdataframe", "subdataframe.jl"),
         ("groupeddataframe", "grouping.jl"),
         ("dataframerow", "dataframerow.jl"),
+        ("dataframerow", "utils.jl"),
 
         ("abstractdataframe", "iteration.jl"),
         ("abstractdataframe", "join.jl"),