--- a/test/REQUIRE
+++ b/test/REQUIRE
@@ -1,1 +1,4 @@
+Compat 0.9.0
 DataStructures
+RDatasets # can be removed when deprecated.jl doesn't test read_rda anymore
+RData