… .travis.yml
… appveyor.yml
… benchmark/datamatrix.jl
… benchmark/datavector.jl
… benchmark/results.csv
… benchmark/runbenchmarks.jl
… docs/mkdocs.yml
… docs/src/index.md
… docs/src/lib/maintypes.md
… docs/src/lib/manipulation.md
… docs/src/lib/utilities.md
… docs/src/man/formulas.md
… docs/src/man/getting_started.md
… docs/src/man/io.md
… docs/src/man/joins.md
… docs/src/man/pooling.md
… docs/src/man/querying_frameworks.md
… docs/src/man/reshaping_and_pivoting.md
… docs/src/man/sorting.md
… docs/src/man/split_apply_combine.md
… docs/src/man/subsets.md
… src/abstractdataframe/iteration.jl
… src/abstractdataframe/join.jl
… src/abstractdataframe/reshape.jl
… src/abstractdataframe/show.jl
… src/abstractdataframe/sort.jl
… src/dataframe/dataframe.jl
… src/dataframe/io.jl
… src/dataframe/sort.jl
… src/dataframerow/dataframerow.jl