PR addresses 4 limitations of current joining and grouping implementation:
- the way row groups are indexed for multicolumn joining/grouping is very sparse (i.e. many usable indices have no rows assigned) and can lead to math overflows even for medium size frames
- rows order in the left-joined or outer-joined frame doesn't match the rows order in the left frame (which is usually nice to maintain)
- for any join kind, including `:inner`, `:semi`, `:anti` full matching of left and right frames is done, which is not necessary
- poor performance for frames containing `PooledDataVector` with large pools due to slow `setindex!(PooledDataArray)`.

Here's the simple benchmarking scripts to test the performance.

``` julia
using Distributions, DataFrames

function random_frame(nrow::Int, col_values::Dict{Symbol, Any})
  DataFrame(Any[isa(col_values[key], PooledDataArray) ?
                @pdata(sample(col_values[key], nrow)) :
                @data(sample(col_values[key], nrow)) for key in keys(col_values)],
            keys(col_values) |> collect)
end

function random_join(kind::Symbol, nrow_left::Int, nrow_right::Int,
                     on_col_values::Dict{Symbol, Any},
                     left_col_values::Dict{Symbol, Any},
                     right_col_values::Dict{Symbol, Any})
  dfl = random_frame(nrow_left, merge(on_col_values, left_col_values))
  dfr = random_frame(nrow_right, merge(on_col_values, right_col_values))
  join(dfl, dfr, on = keys(on_col_values) |> collect, kind = kind)
end

function f(n::Int)
  for i in 1:n
    r = random_join(:outer, 1000, 2000,
                Dict{Symbol,Any}(:A => 1:10, :B => @data([:A, :B, :C, :D]),
                                 :C => 1:10, :D => 1:10),
                Dict{Symbol,Any}(:E => 1:10, :F => @data([:A, :B, :C, :D])),
                Dict{Symbol,Any}(:G => 1:10, :H => @data([:A, :B, :C, :D])))
  end
end

f(1)

@time f(100)
```

For the simple cases PR has almost the same join performance as the current implementation (increased GC overhead is likely due to the allocation of additional arrays that store the rows order in the resulting frame):

Current times:

```
  7.196765 seconds (60.15 M allocations: 1.920 GB, 3.94% gc time)
  7.134183 seconds (60.16 M allocations: 1.920 GB, 4.01% gc time)
  7.180465 seconds (60.15 M allocations: 1.920 GB, 3.95% gc time)
```

PR times:

```
  7.460440 seconds (139.77 M allocations: 3.300 GB, 5.42% gc time)
  7.377374 seconds (139.68 M allocations: 3.298 GB, 5.28% gc time)
  7.340252 seconds (139.88 M allocations: 3.301 GB, 5.60% gc time)
```

However, if in the previous random table generation test, some columns would be converted into  pooled vectors, the current implementation would fail 

``` julia
function g(n::Int)
  for i in 1:n
    r = random_join(:outer, 1000, 2000,
                Dict{Symbol,Any}(:A => 1:10, :B => @data([:A, :B, :C, :D]),
                                 :C => @pdata(1:10), :D => 1:10),
                Dict{Symbol,Any}(:E => 1:10, :F => @pdata([:A, :B, :C, :D])),
                Dict{Symbol,Any}(:G => 1:10, :H => @data([:A, :B, :C, :D])))
  end
end
```

```
julia> g(1)
ERROR: InexactError()
 in setindex! at ./array.jl:303
```

(`InexactError` is a sign that group index exceeded the `typemax(eltype(pooled_vector.refs))`)

While for PR the times are

```
  7.760247 seconds (139.69 M allocations: 3.310 GB, 5.17% gc time)
  7.768788 seconds (139.61 M allocations: 3.308 GB, 5.11% gc time)
  7.776039 seconds (139.76 M allocations: 3.310 GB, 5.18% gc time)
```

The PR wins if the frames contain e.g. `PooledDataVector` columns with suffificiently large pools:

``` julia

function h(n::Int)
    for i in 1:n
        r = random_join(:outer, 10000, 20000,
                       Dict{Symbol,Any}(:A => @pdata(1:10000)),
                       Dict{Symbol,Any}(:B => @pdata(1:10000)),
                       Dict{Symbol,Any}(:C => @pdata(1:10000)))
        end
    end

h(1)

@time h(100)
```

Current implementation:

```
 19.152983 seconds (63.97 M allocations: 2.064 GB, 1.13% gc time)
 19.126810 seconds (63.96 M allocations: 2.064 GB, 1.11% gc time)
 19.198262 seconds (63.97 M allocations: 2.064 GB, 1.02% gc time)
```

PR:

```
  5.145708 seconds (87.02 M allocations: 2.756 GB, 6.64% gc time)
  5.368688 seconds (87.02 M allocations: 2.755 GB, 6.75% gc time)
  5.236543 seconds (87.00 M allocations: 2.741 GB, 6.99% gc time)
```