@alyst 2015-08-26
I've updated the PR. The newer version should be faster and use less memory. Now `_RowGroupDict` implements its own memory-efficient hashing of rows instead of using `Dict{DataFrameRow, Int}`.

As requested by @matthieugomez, here's a small benchmark for `groupby()`
```julia
using RDatasets

diamonds = dataset("ggplot2", "diamonds");
#diamonds[:Cut] = convert(PooledDataArray{ASCIIString, UInt}, diamonds[:Cut])
diamonds[:Cut] = convert(DataArray{ASCIIString}, diamonds[:Cut]);
diamonds[:Clarity] = convert(DataArray{ASCIIString}, diamonds[:Clarity]);
f(n) = for i in 1:n groupby(diamonds, [:Clarity, :Carat]) end;
f(1)
Profile.clear_malloc_data()
@time f(100)
```
(without column `:Cut` conversion `groupby()` fails on master).

`DataFrames.jl` master:
```
2.258086 seconds (37.01 M allocations: 783.709 MB, 2.24% gc time)
```
This PR:
```
  1.833439 seconds (26.86 M allocations: 591.702 MB, 2.00% gc time)
```

In this example the new implementation is both faster and memory efficient. But I don't think the reported numbers reflect the actual efficiency of the implementations. Enormous number of allocations are probably due to elements indexing methods etc. Fixing just JuliaStats/DataArrays.jl#163 already reduced the number of allocations by 50% and there should be some other hotspots. In reality, the difference should be around few hundred allocations.

Note also that doing this on master (after converting resp. columns ref types to `UInt`):
```julia
groupby(diamonds, [:Cut, :Carat, :Clarity, :Depth, :Table, :Price, :Color])
```
could even crash some systems. The problem is that during grouping the estimated number of groups (`ngroups`) would grow very large. The current implementation calls `DataArrays.groupsort_indexer()` to throw away unused group indices, and it tries to allocate vector of `ngroups` length. So we hit the limitations of the current implementation much before overflowing `UInt` (see also #862). One potential fix would be to call `groupsort_indexer()` for each new column added into indexing with the obvious impact on performance.

PS. Travis failure on Julia-nightly is due to JuliaStats/DataArrays.jl#164