@alyst 2015-08-30
For big frames sorting of the row groups can take quite some time. Also, IMHO it's more logical to preserve the original order of rows as much as possible by default. So I've added `sort=` option  (disabled by default) to `by()` and `groupby()`.