@matthieugomez 2015-08-30
Great. dplyr sorts and datatable does not. I like your solution.