@tshort 2015-11-19
There's a lot to like about this PR. The problem is well written out. The added code has good comments and tests. Doc strings are updated. 

My main hesitation is that there's a lot of code churn, and the code looks more complicated. The basic grouping code needs work to improve performance (probably a rewrite). We are at least a factor of ten slower than R's data.table and dplyr packages. This PR doesn't significantly improve grouping speeds based on [this code](https://github.com/JuliaStats/DataFramesMeta.jl/blob/master/test/data.table.timings.jl). My review/opinion here is on the grouping part of the code and not joins.