@alyst 2015-11-20
@tshort Thanks for the review!
I think that to move forward we would also need to add the benchmarks for `getindex()/setindex!()` etc basic functions in DataArrays/NullableArrays (it would also make sense to compare the results with `data.table`/`dplyr`). The suspiciously high number of allocations in the `@time` output (see above) both for master and this PR suggests there could be some type stability problems upstream.