@nalimilan 2016-09-27
Would be interesting to reassess now that we're using `NullableArray`/`CategoricalArray`.