@alyst 2016-10-04
Rebased after #1008.
At the moment relies on JuliaStats/NullableArrays.jl#155 and nalimilan/CategoricalArrays.jl#29