@alyst 2017-01-22
@nalimilan It's not so dead, I've rebased it recently. :) I was just waiting for the `NullableArrays` stuff to settle down until thinking of merging this PR again. It relies on e.g. JuliaStats/NullableArrays.jl#158, which was not merged, because of the anticipated support of lifting in Base.
I'm not up-to-date with the Nullable progress, is it mostly done?