@nalimilan 2017-01-22
> I'm not up-to-date with the Nullable progress, is it mostly done?

AFAIK it's mostly done for a long time now, what's needed is the high-level API (StructuredQueries) to make it easier to use.

Regarding support in NullableArrays, I will comment on the other PR.