@nalimilan 2017-02-18
@alyst The rebase has gone wrong, it's very hard to see your changes now.

Can you tell us more about the groupby algorithm you've adopted here? Cf. discussion at https://github.com/JuliaData/DataTables.jl/issues/3.

Also, what changes do you need in NullableArrays exactly?