@nalimilan 2017-02-18
What happened is that the master branch was moved to DataTables. So if you do `git format-patch` and then `git am` in DataTables, it should work (maybe after replacing `DataFrame` with `DataTable` and `df| with `dt` in the resulting patch).

AFAIK, the current code was inspired by Pandas, except that it didn't implement the code to avoid overflow. Do you think your code is as fast as Pandas, including when all input variables are categorical?