@alyst 2017-02-19
@nalimilan I'm not a Python/Pandas user, so I cannot comment on the performance comparison. It could be that for single-column joins the current code is faster (it's hard to beat, because it just directly uses the column values for indexing), but I spent some time trying to optimize the code for more complex "real life" scenarios. There is a specialized version of column hashing for categorical arrays and nullable arrays. However, to make the joins between nullable/non-nullable and categorical/non-categorical columns work correctly the hashes have to use the underlying values.
It's discussed above that there were some type inference problems in `DataArrays.jl`, so subtle changes in how the values are accessed had big impact on the performance. That was also one of the things I was taking into account, though for NullableArrays it might be different.