@nalimilan 2017-03-06
The PR has now been merged in DataTables (https://github.com/JuliaData/DataTables.jl/pull/17).