@alyst 2017-03-06
JuliaData/DataTables.jl#17 contains many improvements over this PR. So if there is need to introduce these changes to `DataFrames.jl`, JuliaData/DataTables.jl#17 should be used as a reference. Closing.